+++ src/geodef/__init__.py
@@ -18,6 +18,7 @@ from geodef import (
     plot,
     transforms,
     tri,
+    validation,
 )
 from geodef.data import GNSS, DataSet, InSAR, Vertical, spatial_covariance
 from geodef.fault import Fault, magnitude_to_moment, moment_to_magnitude
@@ -58,6 +59,7 @@ __all__ = [
     "plot",
     "transforms",
     "tri",
+    "validation",
     # Data types
     "GNSS",
     "InSAR",